Add initial code for Expo Models
This commit adds the initial code for the Expo Models. It includes the implementation of the `__init__.py` file and the state.json files for the Boolq, copa, Boolq_and_copq models. It also includes the implementation of the `service.proto` file, `Bert.py`, `service_pb2.py`, `CoPA.py`, `Boolq.py`, and `test.ipynb` files.